add env_setting.sh

+++ project/opencv/mat_test/cmake.sh
@@ -1,19 +1,31 @@
 #!/bin/bash
 
-build_dir=`pwd`
+WORK_DIR=`cd "\`dirname "$0"\`"; pwd`
+CMAKE_DIR=${WORK_DIR}/../..
+
+echo "WORK_DIR = ${WORK_DIR} but cur_dir = `pwd`"
+source ${CMAKE_DIR}/env_setting.sh
 PRO_NAME_=mat
 
-rm -rf build
-rm -rf ${PRO_NAME_}
+if [ $# -ge 1 ] && [ $1 == "clean" ]
+then
+    rm -rf build
+    rm -rf ${PRO_NAME_}
+    exit 0
+fi
 
+rm -rf ${WORK_DIR}/build
+rm -rf ${WORK_DIR}/${PRO_NAME_}
+mkdir ${WORK_DIR}/build
+cd ${WORK_DIR}/build
 
-mkdir build
-cd build
-cmake -DPRO_NAME=${PRO_NAME_} -DPRO_DIR=${build_dir} -DENABLE_ZLOG=ON -DENABLE_OPENCV=ON ../../..
-# cmake -DENABLE_ZLOG=ON -DCMAKE_BUILD_TYPE=ON -DCMAKE_EXPORT_COMPILE_COMMANDS=ON -DSRC_EXAM=${build_dir} ../../..
-make
+com_build=-DPRO_NAME=${PRO_NAME_}\ -DPRO_DIR=${WORK_DIR}\ -DPLATFORM=${platform_}\ -DVERSION=${ubuntu_ver_}
+opt_build=-DENABLE_ZLOG=OFF\ -DENABLE_ZMQ_DDS=OFF\ -DENABLE_OPENCV=ON\ -DENABLE_SYS=0FF
+cmake -DENABLE_RELEASE=OFF ${com_build} ${opt_build}  ${CMAKE_DIR}
+
+make -j ${CPU_COUNT_}
 make install
-cd -
-# mv build/${PRO_NAME_} ./
-rm build -rf
+cd ${WORK_DIR}
+rm -rf build
+
 # ulimit -c unlimited
